(feat) move kucoin to new location
