+++ hummingbot/connector/exchange/kucoin/kucoin_auth.py
