+++ hummingbot/connector/exchange/kucoin/kucoin_in_flight_order.pyx
