+++ hummingbot/connector/exchange/kucoin/kucoin_market.pxd
