+++ test/integration/test_kucoin_market.py
@@ -38,7 +38,7 @@ from hummingbot.core.utils.async_utils import (
     safe_ensure_future,
     safe_gather,
 )
-from hummingbot.market.kucoin.kucoin_market import KucoinMarket
+from hummingbot.connector.exchange.kucoin.kucoin_market import KucoinMarket
 from hummingbot.market.market_base import OrderType
 from hummingbot.connector.markets_recorder import MarketsRecorder
 from hummingbot.model.market_state import MarketState
